% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.jsf_env}
\alias{.jsf_env}
\title{Package initialization and JSF module management}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.jsf_env
}
\description{
Internal functions for lazy-loading and caching the Python jsf module
at package startup.
}
\keyword{internal}
