% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ensure_jsf}
\alias{ensure_jsf}
\title{Ensure the Python jsf module is available}
\usage{
ensure_jsf()
}
\value{
Python \code{jsf} module (invisible).
}
\description{
Lazily imports and caches the module.
}
\keyword{internal}
