% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_jsf}
\alias{install_jsf}
\title{Install the Python jsf package}
\usage{
install_jsf(method = "auto", envname = NULL, jsf_path = NULL)
}
\arguments{
\item{method}{Installation method for reticulate::py_install().}

\item{envname}{Target environment name (virtualenv or conda).}

\item{jsf_path}{Optional local path to the jsf package.}
}
\description{
Installs the \code{jsf} Python package into the active Python environment.
}
