% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{jsf_available}
\alias{jsf_available}
\title{Check if the jsf Python module is available}
\usage{
jsf_available()
}
\value{
TRUE if the module can be imported, otherwise FALSE.
}
\description{
Check if the jsf Python module is available
}
