% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsfR-package.R
\docType{package}
\name{jsfR-package}
\alias{jsfR}
\alias{jsfR-package}
\title{jsfR: R Interface to the Jump-Switch-Flow Simulator}
\description{
An R wrapper for the Python \code{jsf} package, providing functions to simulate
and analyze compartmental models using the Jump-Switch-Flow method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DGermano8/jsf}
  \item Report bugs at \url{https://github.com/DGermano8/jsf/issues}
}

}
\author{
Vaibhav Manihar \href{mailto:vaibhav.manihar@gmail.com}{vaibhav.manihar@gmail.com}
}
\keyword{internal}
