% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_error_vs_time}
\alias{plot_error_vs_time}
\title{Plot absolute error between SSA and JSF means over time}
\usage{
plot_error_vs_time(comp)
}
\arguments{
\item{comp}{Comparison data.frame.}
}
\value{
A ggplot object.
}
\description{
Plot absolute error between SSA and JSF means over time
}
