% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot SSA and JSF trajectories with their ensemble means}
\usage{
plot_trajectories(ssa_raw, jsf_raw, comp, lambda, mu, x0)
}
\arguments{
\item{ssa_raw}{Raw SSA data.frame (time, population, rep).}

\item{jsf_raw}{Raw JSF data.frame (time, population, rep).}

\item{comp}{Comparison data.frame from \code{\link[=compute_comparison]{compute_comparison()}}.}

\item{lambda}{Birth rate (for labelling).}

\item{mu}{Death rate (for labelling).}

\item{x0}{Initial population (for labelling).}
}
\value{
A ggplot object.
}
\description{
Plot SSA and JSF trajectories with their ensemble means
}
