% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{print_comparison_summary}
\alias{print_comparison_summary}
\title{Print a summary of comparison metrics}
\usage{
print_comparison_summary(comp)
}
\arguments{
\item{comp}{Data.frame from \code{\link[=compute_comparison]{compute_comparison()}}.}
}
\description{
Print a summary of comparison metrics
}
