% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_jsf}
\alias{theme_jsf}
\title{Plotting functions for JSF simulation results}
\usage{
theme_jsf(base_size = 13)
}
\description{
Functions to visualize trajectories, error heatmaps, and error comparisons
from JSF and SSA simulations.
}
\keyword{Custom}
\keyword{all}
\keyword{for}
\keyword{internal}
\keyword{plots}
\keyword{theme}
